Base

QuanEstimationBase.AD — Method

AD(;max_episode=300, epsilon=0.01, beta1=0.90, beta2=0.99, Adam::Bool=true)

Optimization algorithm: AD.

max_episode: The number of episodes.

epsilon: Learning rate.

beta1: The exponential decay rate for the first moment estimates.

beta2: The exponential decay rate for the second moment estimates.

Adam: Whether or not to use Adam for updating control coefficients.

QuanEstimationBase.CFIM_obj — Method

CFIM_obj(;M=nothing, W=nothing, eps=GLOBAL_EPS)

Choose CFI [$\mathrm{Tr}(WI^{-1})$] as the objective function with $W$ the weight matrix and $I$ the CFIM.

M: A set of positive operator-valued measure (POVM). The default measurement is a set of rank-one symmetric informationally complete POVM (SIC-POVM).

W: Weight matrix.

eps: Machine epsilon.

QuanEstimationBase.CMopt — Type

CMopt(ctrl=nothing, M=nothing, ctrl_bound=[-Inf, Inf], seed=1234)

Control and measurement optimization.

ctrl: Guessed control coefficients.

M: Guessed projective measurement (a set of basis)

ctrl_bound: Lower and upper bounds of the control coefficients.

seed: Random seed.

QuanEstimationBase.ControlOpt — Method

ControlOpt(ctrl=nothing, ctrl_bound=[-Inf, Inf], seed=1234)

Control optimization.

ctrl: Guessed control coefficients.

ctrl_bound: Lower and upper bounds of the control coefficients.

seed: Random seed.

QuanEstimationBase.DDPG — Method

DDPG(;max_episode::Int=500, layer_num::Int=3, layer_dim::Int=200, seed::Number=1234)

Optimization algorithm: DE.

max_episode: The number of populations.

layer_num: The number of layers (include the input and output layer).

layer_dim: The number of neurons in the hidden layer.

seed: Random seed.

QuanEstimationBase.DE — Method

DE(;max_episode::Number=1000, p_num::Number=10, ini_population=nothing, c::Number=1.0, cr::Number=0.5, seed::Number=1234)

Optimization algorithm: DE.

max_episode: The number of populations.

p_num: The number of particles.

ini_population: Initial guesses of the optimization variables.

c: Mutation constant.

cr: Crossover constant.

QuanEstimationBase.GRAPE — Method

GRAPE(;max_episode=300, epsilon=0.01, beta1=0.90, beta2=0.99, Adam::Bool=true)

Control optimization algorithm: GRAPE.

max_episode: The number of episodes.

epsilon: Learning rate.

beta1: The exponential decay rate for the first moment estimates.

beta2: The exponential decay rate for the second moment estimates.

Adam: Whether or not to use Adam for updating control coefficients.

QuanEstimationBase.HCRB_obj — Method

HCRB_obj(;W=nothing, eps=GLOBAL_EPS)

Choose HCRB as the objective function.

W: Weight matrix.

eps: Machine epsilon.

QuanEstimationBase.LLD — Method

LLD(ρ::Matrix{T}, dρ::Vector{Matrix{T}}; rep="original", eps=GLOBAL_EPS) where {T<:Complex}

Calculate the left logarrithmic derivatives (LLDs). The LLD operator is defined as $\partial_{a}\rho=\mathcal{R}_a^{\dagger}\rho$, where ρ is the parameterized density matrix.

ρ: Density matrix.

dρ: Derivatives of the density matrix with respect to the unknown parameters to be estimated. For example, drho[1] is the derivative vector with respect to the first parameter.

rep: Representation of the LLD operator. Options can be: "original" (default) and "eigen".

eps: Machine epsilon.

QuanEstimationBase.LLD — Method

LLD(ρ::Matrix{T}, dρ::Matrix{T}; rep="original", eps=GLOBAL_EPS) where {T<:Complex}

When applied to the case of single parameter.

QuanEstimationBase.NM — Method

NM(;max_episode::Int=1000, p_num::Int=10, nelder_mead=nothing, ar::Number=1.0, ae::Number=2.0, ac::Number=0.5, as0::Number=0.5, seed::Number=1234)

State optimization algorithm: NM.

max_episode: The number of populations.

p_num: The number of the input states.

nelder_mead: Initial guesses of the optimization variables.

ar: Reflection constant.

ae: Expansion constant.

ac: Constraction constant.

as0: Shrink constant.

QuanEstimationBase.PSO — Method

PSO(;max_episode::Union{T,Vector{T}} where {T<:Int}=[1000, 100], p_num::Number=10, ini_particle=nothing, c0::Number=1.0, c1::Number=2.0, c2::Number=2.0, seed::Number=1234)

Optimization algorithm: PSO.

max_episode: The number of episodes, it accepts both integer and array with two elements.

p_num: The number of particles.

ini_particle: Initial guesses of the optimization variables.

c0: The damping factor that assists convergence, also known as inertia weight.

c1: The exploitation weight that attracts the particle to its best previous position, also known as cognitive learning factor.

c2: The exploitation weight that attracts the particle to the best position in the neighborhood, also known as social learning factor.

QuanEstimationBase.QFIM_obj — Method

QFIM_obj(;W=nothing, eps=GLOBAL_EPS, LDtype::Symbol=:SLD)

Choose QFI [$\mathrm{Tr}(WF^{-1})$] as the objective function with $W$ the weight matrix and $F$ the QFIM.

W: Weight matrix.

eps: Machine epsilon.

LDtype: Types of QFI (QFIM) can be set as the objective function. Options are :SLD (default), :RLD and :LLD.

QuanEstimationBase.RI — Method

RI(;max_episode::Int=300, seed::Number=1234)

State optimization algorithm: RI.

max_episode: The number of episodes.

QuanEstimationBase.RLD — Method

RLD(ρ::Matrix{T}, dρ::Vector{Matrix{T}}; rep="original", eps=GLOBAL_EPS) where {T<:Complex}

Calculate the right logarrithmic derivatives (RLDs). The RLD operator is defined as $\partial_{a}\rho=\rho \mathcal{R}_a$, where $\rho$ is the parameterized density matrix.

ρ: Density matrix.

dρ: Derivatives of the density matrix with respect to the unknown parameters to be estimated. For example, drho[1] is the derivative vector with respect to the first parameter.

rep: Representation of the RLD operator. Options can be: "original" (default) and "eigen".

eps: Machine epsilon.

QuanEstimationBase.RLD — Method

RLD(ρ::Matrix{T}, dρ::Matrix{T}; rep="original", eps=GLOBAL_EPS) where {T<:Complex}

When applied to the case of single parameter.

QuanEstimationBase.SCMopt — Type

SCMopt(psi=nothing, ctrl=nothing, M=nothing, ctrl_bound=[-Inf, Inf], seed=1234)

State, control and measurement optimization.

psi: Guessed probe state.

ctrl: Guessed control coefficients.

M: Guessed projective measurement (a set of basis).

ctrl_bound: Lower and upper bounds of the control coefficients.

seed: Random seed.

QuanEstimationBase.SCopt — Type

SCopt(psi=nothing, ctrl=nothing, ctrl_bound=[-Inf, Inf], seed=1234)

State and control optimization.

psi: Guessed probe state.

ctrl: Guessed control coefficients.

ctrl_bound: Lower and upper bounds of the control coefficients.

seed: Random seed.

QuanEstimationBase.SLD — Method

SLD(ρ::Matrix{T}, dρ::Vector{Matrix{T}}; rep="original", eps=GLOBAL_EPS) where {T<:Complex}

Calculate the symmetric logarrithmic derivatives (SLDs). The SLD operator $L_a$ is defined as$\partial_{a}\rho=\frac{1}{2}(\rho L_{a}+L_{a}\rho)$, where $\rho$ is the parameterized density matrix.

ρ: Density matrix.

dρ: Derivatives of the density matrix with respect to the unknown parameters to be estimated. For example, drho[1] is the derivative vector with respect to the first parameter.

rep: Representation of the SLD operator. Options can be: "original" (default) and "eigen" .

eps: Machine epsilon.

QuanEstimationBase.SLD — Method

SLD(ρ::Matrix{T}, dρ::Matrix{T}; rep="original", eps=GLOBAL_EPS) where {T<:Complex}

When applied to the case of single parameter.

QuanEstimationBase.SMopt — Type

SMopt(psi=nothing, M=nothing, seed=1234)

State and control optimization.

psi: Guessed probe state.

M: Guessed projective measurement (a set of basis).

seed: Random seed.

QuanEstimationBase.StateOpt — Method

StateOpt(psi=nothing, seed=1234)

State optimization.

psi: Guessed probe state.

seed: Random seed.

QuanEstimationBase.autoGRAPE — Method

autoGRAPE(;max_episode=300, epsilon=0.01, beta1=0.90, beta2=0.99, Adam::Bool=true)

Control optimization algorithm: auto-GRAPE.

max_episode: The number of episodes.

epsilon: Learning rate.

beta1: The exponential decay rate for the first moment estimates.

beta2: The exponential decay rate for the second moment estimates.

Adam: Whether or not to use Adam for updating control coefficients.

QuanEstimationBase.BCB — Method

BCB(x, p, rho; W=nothing, eps=GLOBAL_EPS)

Calculation of the minimum Bayesian cost with a quadratic cost function.

x: The regimes of the parameters for the integral.

p: The prior distribution.

rho: Parameterized density matrix.

W: Weight matrix.

eps: Machine epsilon.

QuanEstimationBase.BCFIM — Method

BCFIM(x::AbstractVector, p, rho, drho; M=nothing, eps=GLOBAL_EPS)

Calculation of the Bayesian classical Fisher information (BCFI) and the Bayesian classical Fisher information matrix (BCFIM) of the form $\mathcal{I}_{\mathrm{Bayes}}=\int p(\textbf{x})\mathcal{I}\mathrm{d}\textbf{x}$ with $\mathcal{I}$ the CFIM and $p(\textbf{x})$ the prior distribution.

x: The regimes of the parameters for the integral.

p: The prior distribution.

rho: Parameterized density matrix.

drho: Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

M: A set of positive operator-valued measure (POVM). The default measurement is a set of rank-one symmetric informationally complete POVM (SIC-POVM).

eps: Machine epsilon.

QuanEstimationBase.BCRB — Method

BCRB(x::AbstractVector, p, dp, rho, drho; M=nothing, b=nothing, db=nothing, btype=1, eps=GLOBAL_EPS)

Calculation of the Bayesian Cramer-Rao bound (BCRB).

x: The regimes of the parameters for the integral.

p: The prior distribution.

dp: Derivatives of the prior distribution with respect to the unknown parameters to be estimated. For example, dp[0] is the derivative vector on the first parameter.

rho: Parameterized density matrix.

drho: Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

M: A set of positive operator-valued measure (POVM). The default measurement is a set of rank-one symmetric informationally complete POVM (SIC-POVM).

b: Vector of biases of the form $\textbf{b}=(b(x_0),b(x_1),\dots)^{\mathrm{T}}$.

db: Derivatives of b on the unknown parameters to be estimated, It should be expressed as $\textbf{b}'=(\partial_0 b(x_0),\partial_1 b(x_1),\dots)^{\mathrm{T}}$.

btype: Types of the BCRB. Options are 1, 2 and 3.

eps: Machine epsilon.

QuanEstimationBase.BQCRB — Method

BQCRB(x::AbstractVector, p, dp, rho, drho; b=nothing, db=nothing, LDtype=:SLD, btype=1, eps=GLOBAL_EPS)

Calculation of the Bayesian quantum Cramer-Rao bound (BQCRB).

x: The regimes of the parameters for the integral.

p: The prior distribution.

dp: Derivatives of the prior distribution with respect to the unknown parameters to be estimated. For example, dp[0] is the derivative vector on the first parameter.

rho: Parameterized density matrix.

drho: Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

b: Vector of biases of the form $\textbf{b}=(b(x_0),b(x_1),\dots)^{\mathrm{T}}$.

db: Derivatives of b on the unknown parameters to be estimated, It should be expressed as $\textbf{b}'=(\partial_0 b(x_0),\partial_1 b(x_1),\dots)^{\mathrm{T}}$.

LDtype: Types of QFI (QFIM) can be set as the objective function. Options are "SLD" (default), "RLD" and "LLD".

btype: Types of the BCRB. Options are 1, 2 and 3.

eps: Machine epsilon.

QuanEstimationBase.BQFIM — Method

BQFIM(x::AbstractVector, p, rho, drho; LDtype=:SLD, eps=GLOBAL_EPS)

Calculation of the Bayesian quantum Fisher information (BQFI) and the Bayesian quantum Fisher information matrix (BQFIM) of the form $\mathcal{F}_{\mathrm{Bayes}}=\int p(\textbf{x})\mathcal{F}\mathrm{d}\textbf{x}$ with $\mathcal{F}$ the QFIM of all types and $p(\textbf{x})$ the prior distribution.

x: The regimes of the parameters for the integral.

p: The prior distribution.

rho: Parameterized density matrix.

drho: Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

LDtype: Types of QFI (QFIM) can be set as the objective function. Options are "SLD" (default), "RLD" and "LLD".

eps: Machine epsilon.

QuanEstimationBase.Bayes — Method

Bayes(x, p, rho, y; M=nothing, savefile=false)

Bayesian estimation. The prior distribution is updated via the posterior distribution obtained by the Bayes’ rule and the estimated value of parameters obtained via the maximum a posteriori probability (MAP).

x: The regimes of the parameters for the integral.

p: The prior distribution.

rho: Parameterized density matrix.

y: The experimental results obtained in practice.

M: A set of positive operator-valued measure (POVM). The default measurement is a set of rank-one symmetric informationally complete POVM (SIC-POVM).

savefile: Whether or not to save all the posterior distributions.

QuanEstimationBase.BayesCost — Method

BayesCost(x, p, xest, rho, M; W=nothing, eps=GLOBAL_EPS)

Calculation of the average Bayesian cost with a quadratic cost function.

x: The regimes of the parameters for the integral.

p: The prior distribution.

xest: The estimators.

rho: Parameterized density matrix.

M: A set of POVM.

W: Weight matrix.

eps: Machine epsilon.

QuanEstimationBase.CFIM — Method

CFIM(ρ::Matrix{T}, dρ::Vector{Matrix{T}}, M; eps=GLOBAL_EPS) where {T<:Complex}

Calculate the classical Fisher information matrix (CFIM).

ρ: Density matrix.

dρ: Derivatives of the density matrix with respect to the unknown parameters to be estimated. For example, drho[1] is the derivative vector with respect to the first parameter.

M: A set of positive operator-valued measure (POVM). The default measurement is a set of rank-one symmetric informationally complete POVM (SIC-POVM).

eps: Machine epsilon.

QuanEstimationBase.CFIM — Method

CFIM(ρ::Matrix{T}, dρ::Vector{Matrix{T}}; M=nothing, eps=GLOBAL_EPS) where {T<:Complex}

When the set of POVM is not given. Calculate the CFIM with SIC-POVM. The SIC-POVM is generated from the Weyl-Heisenberg covariant SIC-POVM fiducial state which can be downloaded from here.

QuanEstimationBase.CFIM — Method

CFIM(ρ::Matrix{T}, dρ::Matrix{T}, M; eps=GLOBAL_EPS) where {T<:Complex}

When applied to the case of single parameter. Calculate the classical Fisher information (CFI).

QuanEstimationBase.CFIM — Method

CFIM(ρ::Matrix{T}, dρ::Matrix{T}; eps=GLOBAL_EPS) where {T<:Complex}

When applied to the case of single parameter and the set of POVM is not given. Calculate the CFI with SIC-POVM.

QuanEstimationBase.FIM — Method

FIM(p::Vector{R}, dp::Vector{R}; eps=GLOBAL_EPS) where {R<:Real}

Calculation of the classical Fisher information matrix for classical scenarios.

p: The probability distribution.

dp: Derivatives of the probability distribution on the unknown parameters to be estimated. For example, dp[0] is the derivative vector on the first parameter.

eps: Machine epsilon.

QuanEstimationBase.FIM — Method

FIM(p::Vector{R}, dp::Vector{R}; eps=GLOBAL_EPS) where {R<:Real}

When applied to the case of single parameter and the set of POVM is not given. Calculate the classical Fisher information for classical scenarios.

QuanEstimationBase.FI_Expt — Method

FI_Expt(y1, y2, dx; ftype=:norm)

Calculate the classical Fisher information (CFI) based on the experiment data.

y1: Experimental data obtained at the truth value (x).

y1: Experimental data obtained at x+dx.

dx: A known small drift of the parameter.

ftype: The distribution the data follows. Options are: norm, gamma, rayleigh, and poisson.

QuanEstimationBase.HCRB — Method

HCRB(ρ::AbstractMatrix, dρ::AbstractVector, W::AbstractMatrix; eps=GLOBAL_EPS)

Caltulate the Holevo Cramer-Rao bound (HCRB) via the semidefinite program (SDP).

ρ: Density matrix.

dρ: Derivatives of the density matrix on the unknown parameters to be estimated. For example, drho[0] is the derivative vector on the first parameter.

W: Weight matrix.

eps: Machine epsilon.

QuanEstimationBase.MLE — Method

MLE(x, rho, y; M=nothing, savefile=false)

Bayesian estimation. The estimated value of parameters obtained via the maximum likelihood estimation (MLE).

x: The regimes of the parameters for the integral.

rho: Parameterized density matrix.

y: The experimental results obtained in practice.

M: A set of positive operator-valued measure (POVM). The default measurement is a set of rank-one symmetric informationally complete POVM (SIC-POVM).

savefile: Whether or not to save all the posterior distributions.

QuanEstimationBase.MeasurementOpt — Method

MeasurementOpt(mtype=:Projection, kwargs...)

Measurement optimization.

mtype: The type of scenarios for the measurement optimization. Options are :Projection (default), :LC and :Rotation.

kwargs...: keywords and the correponding default vaules. mtype=:Projection, mtype=:LC and mtype=:Rotation, the kwargs... are M=nothing, B=nothing, POVM_basis=nothing, and s=nothing, POVM_basis=nothing, respectively.

QuanEstimationBase.NHB — Method

NHB(ρ::AbstractMatrix, dρ::AbstractVector, W::AbstractMatrix)

Nagaoka-Hayashi bound (NHB) via the semidefinite program (SDP).

ρ: Density matrix.

dρ: Derivatives of the density matrix on the unknown parameters to be estimated. For example, drho[0] is the derivative vector on the first parameter.

W: Weight matrix.

QuanEstimationBase.OBB — Method

OBB(x::AbstractVector, p, dp, rho, drho, d2rho; LDtype=:SLD, eps=GLOBAL_EPS)

Calculation of the Bayesian version of Cramer-Rao bound introduced by Van Trees (VTB).

x: The regimes of the parameters for the integral.

p: The prior distribution.

dp: Derivatives of the prior distribution with respect to the unknown parameters to be estimated. For example, dp[0] is the derivative vector on the first parameter.

rho: Parameterized density matrix.

drho: Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

d2rho: Second order Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

LDtype: Types of QFI (QFIM) can be set as the objective function. Options are "SLD" (default), "RLD" and "LLD".

eps: Machine epsilon.

QuanEstimationBase.QFIM — Method

QFIM(ρ::Matrix{T}, dρ::Vector{Matrix{T}}; LDtype=:SLD, exportLD::Bool= false, eps=GLOBAL_EPS) where {T<:Complex}

When applied to the case of single parameter. Calculation of the quantum Fisher information (QFI) for all types.

QuanEstimationBase.QFIM — Method

QFIM(ρ::Matrix{T}, dρ::Matrix{T}; LDtype=:SLD, exportLD::Bool= false, eps=GLOBAL_EPS) where {T<:Complex}

Calculation of the quantum Fisher information (QFI) for all types.

ρ: Density matrix.

dρ: Derivatives of the density matrix with respect to the unknown parameters to be estimated. For example, drho[1] is the derivative vector with respect to the first parameter.

LDtype: Types of QFI (QFIM) can be set as the objective function. Options are :SLD (default), :RLD and :LLD.

exportLD: export logarithmic derivatives apart from F.

eps: Machine epsilon.

QuanEstimationBase.QFIM_Gauss — Method

QFIM_Gauss(R̄::V, dR̄::VV, D::M, dD::VM) where {V,VV,M,VM<:AbstractVecOrMat}

Calculate the SLD based quantum Fisher information matrix (QFIM) with gaussian states.

R̄ : First-order moment.

dR̄: Derivatives of the first-order moment with respect to the unknown parameters to be estimated. For example, dR[1] is the derivative vector on the first parameter.

D: Second-order moment.

dD: Derivatives of the second-order moment with respect to the unknown parameters to be estimated.

eps: Machine epsilon.

QuanEstimationBase.QFIM_Kraus — Method

QFIM_Kraus(ρ0::AbstractMatrix, K::AbstractVector, dK::AbstractVector; LDtype=:SLD, exportLD::Bool=false, eps=GLOBAL_EPS)

Calculation of the quantum Fisher information (QFI) and quantum Fisher information matrix (QFIM) with Kraus operator(s) for all types.

ρ0: Density matrix.

K: Kraus operator(s).

dK: Derivatives of the Kraus operator(s) on the unknown parameters to be estimated. For example, dK[0] is the derivative vector on the first parameter.

LDtype: Types of QFI (QFIM) can be set as the objective function. Options are :SLD (default), :RLD and :LLD.

exportLD: Whether or not to export the values of logarithmic derivatives. If set True then the the values of logarithmic derivatives will be exported.

eps: Machine epsilon.

QuanEstimationBase.QVTB — Method

QVTB(x::AbstractVector, p, dp, rho, drho; LDtype=:SLD, eps=GLOBAL_EPS)

Calculation of the Bayesian version of Cramer-Rao bound in troduced by Van Trees (VTB).

x: The regimes of the parameters for the integral.

p: The prior distribution.

dp: Derivatives of the prior distribution with respect to the unknown parameters to be estimated. For example, dp[0] is the derivative vector on the first parameter.

rho: Parameterized density matrix.

drho: Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

LDtype: Types of QFI (QFIM) can be set as the objective function. Options are "SLD" (default), "RLD" and "LLD".

eps: Machine epsilon.

QuanEstimationBase.QZZB — Method

QZZB(x::AbstractVector, p::AbstractVector, rho::AbstractVecOrMat; eps=GLOBAL_EPS)

Calculation of the quantum Ziv-Zakai bound (QZZB).

x: The regimes of the parameters for the integral.

p: The prior distribution.

rho: Parameterized density matrix.

eps: Machine epsilon.

QuanEstimationBase.SIC — Method

SIC(dim::Int64)

Generation of a set of rank-one symmetric informationally complete positive operator-valued measure (SIC-POVM).

dim: The dimension of the system.

Note: SIC-POVM is calculated by the Weyl-Heisenberg covariant SIC-POVM fiducial state which can be downloaded from here.

QuanEstimationBase.SpinSqueezing — Method

SpinSqueezing(ρ::AbstractMatrix; basis="Dicke", output="KU")

Calculate the spin squeezing parameter for the input density matrix. The basis can be "Dicke" for the Dicke basis, or "Pauli" for the Pauli basis. The output can be both "KU"(for spin squeezing defined by Kitagawa and Ueda) and "WBIMH"(for spin squeezing defined by Wineland et al.).

QuanEstimationBase.TargetTime — Method

TargetTime(f::Number, tspan::AbstractVector, func::Function, args...; kwargs...)

Calculate the minimum time to reach a precision limit of given level. The func can be any objective function during the control optimization, e.g. QFIM, CFIM, HCRB, etc.

QuanEstimationBase.VTB — Method

VTB(x::AbstractVector, p, dp, rho, drho; M=nothing, eps=GLOBAL_EPS)

Calculation of the Bayesian version of Cramer-Rao bound introduced by Van Trees (VTB).

x: The regimes of the parameters for the integral.

p: The prior distribution.

dp: Derivatives of the prior distribution with respect to the unknown parameters to be estimated. For example, dp[0] is the derivative vector on the first parameter.

rho: Parameterized density matrix.

drho: Derivatives of the parameterized density matrix (rho) with respect to the unknown parameters to be estimated.

M: A set of positive operator-valued measure (POVM). The default measurement is a set of rank-one symmetric informationally complete POVM (SIC-POVM).

eps: Machine epsilon.

QuanEstimationBase.evolve — Method

evolve(dynamics::Kraus{dm})

Evolution of density matrix under time-independent Hamiltonian without noise and controls.

QuanEstimationBase.evolve — Method

evolve(dynamics::Kraus{ket})

Evolution of pure states under time-independent Hamiltonian without noise and controls

QuanEstimationBase.expm — Method

expm(tspan::AbstractVector, ρ0::AbstractMatrix, H0::AbstractMatrix, dH::AbstractMatrix; decay::Union{AbstractVector, Nothing}=nothing, Hc::Union{AbstractVector, Nothing}=nothing, ctrl::Union{AbstractVector, Nothing}=nothing)

When applied to the case of single parameter.

QuanEstimationBase.expm — Method

expm(tspan::AbstractVector, ρ0::AbstractMatrix, H0::AbstractMatrix, dH::AbstractVector; decay::Union{AbstractVector, Nothing}=nothing, Hc::Union{AbstractVector, Nothing}=nothing, ctrl::Union{AbstractVector, Nothing}=nothing)

The dynamics of a density matrix is of the form $\partial_t\rho=-i[H,\rho]+\sum_i \gamma_i\left(\Gamma_i\rho\Gamma^{\dagger}_i-\frac{1}{2}\left\{\rho,\Gamma^{\dagger}_i \Gamma_i \right\}\right)$, where $\rho$ is the evolved density matrix, $H$ is the Hamiltonian of the system, $\Gamma_i$ and $\gamma_i$ are the $i\mathrm{th}$ decay operator and the corresponding decay rate.

tspan: Time length for the evolution.

ρ0: Initial state (density matrix).

H0: Free Hamiltonian.

dH: Derivatives of the free Hamiltonian with respect to the unknown parameters to be estimated. For example, dH[0] is the derivative vector on the first parameter.

decay: Decay operators and the corresponding decay rates. Its input rule is decay=[[$\Gamma_1$, $\gamma_1$], [$\Gamma_2$, $\gamma_2$],...], where $\Gamma_1$ $(\Gamma_2)$ represents the decay operator and $\gamma_1$ $(\gamma_2)$ is the corresponding decay rate.

Hc: Control Hamiltonians.

ctrl: Control coefficients.

QuanEstimationBase.ode — Method

ode(tspan::AbstractVector, ρ0::AbstractMatrix, H0::AbstractMatrix, dH::AbstractMatrix; decay::Union{AbstractVector, Nothing}=nothing, Hc::Union{AbstractVector, Nothing}=nothing, ctrl::Union{AbstractVector, Nothing}=nothing)

When applied to the case of single parameter.

QuanEstimationBase.ode — Method

ode(tspan::AbstractVector, ρ0::AbstractMatrix, H0::AbstractVector, dH::AbstractVector; decay::Union{AbstractVector, Nothing}=nothing, Hc::Union{AbstractVector, Nothing}=nothing, ctrl::Union{AbstractVector, Nothing}=nothing)

The dynamics of a density matrix is of the form $\partial_t\rho=-i[H,\rho]+\sum_i \gamma_i\left(\Gamma_i\rho\Gamma^{\dagger}_i-\frac{1}{2}\left\{\rho,\Gamma^{\dagger}_i \Gamma_i \right\}\right)$, where $\rho$ is the evolved density matrix, $H$ is the Hamiltonian of the system, $\Gamma_i$ and $\gamma_i$ are the $i\mathrm{th}$ decay operator and the corresponding decay rate.

tspan: Time length for the evolution.

ρ0: Initial state (density matrix).

H0: Free Hamiltonian.

dH: Derivatives of the free Hamiltonian with respect to the unknown parameters to be estimated. For example, dH[0] is the derivative vector on the first parameter.

decay: Decay operators and the corresponding decay rates. Its input rule is decay=[[$\Gamma_1$, $\gamma_1$], [$\Gamma_2$, $\gamma_2$],...], where $\Gamma_1$ $(\Gamma_2)$ represents the decay operator and $\gamma_1$ $(\gamma_2)$ is the corresponding decay rate.

Hc: Control Hamiltonians.

ctrl: Control coefficients.

QuanEstimationBase.offline — Method

offline(apt::Adapt_MZI, alg; target::Symbol=:sharpness, eps = GLOBAL_EPS, seed=1234)

Offline adaptive phase estimation in the MZI.

apt: Adaptive MZI struct which contains x, p, and rho0.

alg: The algorithms for searching the optimal tunable phase. Here, DE and PSO are available.

target: Setting the target function for calculating the tunable phase. Options are: "sharpness" and "MI".

eps: Machine epsilon.

seed: Random seed.

QuanEstimationBase.online — Method

online(apt::Adapt_MZI; target::Symbol=:sharpness, output::String="phi")

Online adaptive phase estimation in the MZI.

apt: Adaptive MZI struct which contains x, p, and rho0.

target: Setting the target function for calculating the tunable phase. Options are: "sharpness" and "MI".

output: Choose the output variables. Options are: "phi" and "dphi".

QuanEstimationBase.suN_generator — Method

suN_generator(n::Int64)

Generation of the SU($N$) generators with $N$ the dimension of the system.

N: The dimension of the system.